amd64_warnings * changed a few ints to long that were used as pointers or vv * this should reduce warnings on 64-bit systems
git-svn-id: https://svn.parrot.org/parrot/trunk@5510 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ include/parrot/oplib.h
@@ -41,7 +41,7 @@ typedef enum {
 } op_lib_flags_enum;
 
 /* when init = true initialize, else de_initialize */
-typedef op_lib_t *(*oplib_init_f)(int init);
+typedef op_lib_t *(*oplib_init_f)(long init);
 
 /* core.ops special opcode numbers */
 typedef enum {
